move legacy settings API to "internal"
