+++ clients/python/lakefs_client/api/repositories_api.py
@@ -22,9 +22,7 @@ from lakefs_client.model_utils import (  # noqa: F401
     none_type,
     validate_and_convert_types
 )
-from lakefs_client.model.branch_protection_rule import BranchProtectionRule
 from lakefs_client.model.error import Error
-from lakefs_client.model.inline_object1 import InlineObject1
 from lakefs_client.model.repository import Repository
 from lakefs_client.model.repository_creation import RepositoryCreation
 from lakefs_client.model.repository_list import RepositoryList
@@ -42,68 +40,6 @@ class RepositoriesApi(object):
         if api_client is None:
             api_client = ApiClient()
         self.api_client = api_client
-        self.create_branch_protection_rule_endpoint = _Endpoint(
-            settings={
-                'response_type': None,
-                'auth': [
-                    'basic_auth',
-                    'cookie_auth',
-                    'jwt_token',
-                    'oidc_auth',
-                    'saml_auth'
-                ],
-                'endpoint_path': '/repositories/{repository}/branch_protection',
-                'operation_id': 'create_branch_protection_rule',
-                'http_method': 'POST',
-                'servers': None,
-            },
-            params_map={
-                'all': [
-                    'repository',
-                    'branch_protection_rule',
-                ],
-                'required': [
-                    'repository',
-                    'branch_protection_rule',
-                ],
-                'nullable': [
-                ],
-                'enum': [
-                ],
-                'validation': [
-                ]
-            },
-            root_map={
-                'validations': {
-                },
-                'allowed_values': {
-                },
-                'openapi_types': {
-                    'repository':
-                        (str,),
-                    'branch_protection_rule':
-                        (BranchProtectionRule,),
-                },
-                'attribute_map': {
-                    'repository': 'repository',
-                },
-                'location_map': {
-                    'repository': 'path',
-                    'branch_protection_rule': 'body',
-                },
-                'collection_format_map': {
-                }
-            },
-            headers_map={
-                'accept': [
-                    'application/json'
-                ],
-                'content_type': [
-                    'application/json'
-                ]
-            },
-            api_client=api_client
-        )
         self.create_repository_endpoint = _Endpoint(
             settings={
                 'response_type': (Repository,),
@@ -165,68 +101,6 @@ class RepositoriesApi(object):
             },
             api_client=api_client
         )
-        self.delete_branch_protection_rule_endpoint = _Endpoint(
-            settings={
-                'response_type': None,
-                'auth': [
-                    'basic_auth',
-                    'cookie_auth',
-                    'jwt_token',
-                    'oidc_auth',
-                    'saml_auth'
-                ],
-                'endpoint_path': '/repositories/{repository}/branch_protection',
-                'operation_id': 'delete_branch_protection_rule',
-                'http_method': 'DELETE',
-                'servers': None,
-            },
-            params_map={
-                'all': [
-                    'repository',
-                    'inline_object1',
-                ],
-                'required': [
-                    'repository',
-                    'inline_object1',
-                ],
-                'nullable': [
-                ],
-                'enum': [
-                ],
-                'validation': [
-                ]
-            },
-            root_map={
-                'validations': {
-                },
-                'allowed_values': {
-                },
-                'openapi_types': {
-                    'repository':
-                        (str,),
-                    'inline_object1':
-                        (InlineObject1,),
-                },
-                'attribute_map': {
-                    'repository': 'repository',
-                },
-                'location_map': {
-                    'repository': 'path',
-                    'inline_object1': 'body',
-                },
-                'collection_format_map': {
-                }
-            },
-            headers_map={
-                'accept': [
-                    'application/json'
-                ],
-                'content_type': [
-                    'application/json'
-                ]
-            },
-            api_client=api_client
-        )
         self.delete_repository_endpoint = _Endpoint(
             settings={
                 'response_type': None,
@@ -240,61 +114,6 @@ class RepositoriesApi(object):
                 'endpoint_path': '/repositories/{repository}',
                 'operation_id': 'delete_repository',
                 'http_method': 'DELETE',
-                'servers': None,
-            },
-            params_map={
-                'all': [
-                    'repository',
-                ],
-                'required': [
-                    'repository',
-                ],
-                'nullable': [
-                ],
-                'enum': [
-                ],
-                'validation': [
-                ]
-            },
-            root_map={
-                'validations': {
-                },
-                'allowed_values': {
-                },
-                'openapi_types': {
-                    'repository':
-                        (str,),
-                },
-                'attribute_map': {
-                    'repository': 'repository',
-                },
-                'location_map': {
-                    'repository': 'path',
-                },
-                'collection_format_map': {
-                }
-            },
-            headers_map={
-                'accept': [
-                    'application/json'
-                ],
-                'content_type': [],
-            },
-            api_client=api_client
-        )
-        self.get_branch_protection_rules_endpoint = _Endpoint(
-            settings={
-                'response_type': ([BranchProtectionRule],),
-                'auth': [
-                    'basic_auth',
-                    'cookie_auth',
-                    'jwt_token',
-                    'oidc_auth',
-                    'saml_auth'
-                ],
-                'endpoint_path': '/repositories/{repository}/branch_protection',
-                'operation_id': 'get_branch_protection_rules',
-                'http_method': 'GET',
                 'servers': None,
             },
             params_map={
@@ -517,75 +336,6 @@ class RepositoriesApi(object):
             api_client=api_client
         )
 
-    def create_branch_protection_rule(
-        self,
-        repository,
-        branch_protection_rule,
-        **kwargs
-    ):
-        """create_branch_protection_rule  # noqa: E501
-
-        This method makes a synchronous HTTP request by default. To make an
-        asynchronous HTTP request, please pass async_req=True
-
-        >>> thread = api.create_branch_protection_rule(repository, branch_protection_rule, async_req=True)
-        >>> result = thread.get()
-
-        Args:
-            repository (str):
-            branch_protection_rule (BranchProtectionRule):
-
-        Keyword Args:
-            _return_http_data_only (bool): response data without head status
-                code and headers. Default is True.
-            _preload_content (bool): if False, the urllib3.HTTPResponse object
-                will be returned without reading/decoding response data.
-                Default is True.
-            _request_timeout (int/float/tuple): timeout setting for this request. If
-                one number provided, it will be total request timeout. It can also
-                be a pair (tuple) of (connection, read) timeouts.
-                Default is None.
-            _check_input_type (bool): specifies if type checking
-                should be done one the data sent to the server.
-                Default is True.
-            _check_return_type (bool): specifies if type checking
-                should be done one the data received from the server.
-                Default is True.
-            _host_index (int/None): specifies the index of the server
-                that we want to use.
-                Default is read from the configuration.
-            async_req (bool): execute request asynchronously
-
-        Returns:
-            None
-                If the method is called asynchronously, returns the request
-                thread.
-        """
-        kwargs['async_req'] = kwargs.get(
-            'async_req', False
-        )
-        kwargs['_return_http_data_only'] = kwargs.get(
-            '_return_http_data_only', True
-        )
-        kwargs['_preload_content'] = kwargs.get(
-            '_preload_content', True
-        )
-        kwargs['_request_timeout'] = kwargs.get(
-            '_request_timeout', None
-        )
-        kwargs['_check_input_type'] = kwargs.get(
-            '_check_input_type', True
-        )
-        kwargs['_check_return_type'] = kwargs.get(
-            '_check_return_type', True
-        )
-        kwargs['_host_index'] = kwargs.get('_host_index')
-        kwargs['repository'] = \
-            repository
-        kwargs['branch_protection_rule'] = \
-            branch_protection_rule
-        return self.create_branch_protection_rule_endpoint.call_with_http_info(**kwargs)
-
     def create_repository(
         self,
         repository_creation,
@@ -652,75 +402,6 @@ class RepositoriesApi(object):
             repository_creation
         return self.create_repository_endpoint.call_with_http_info(**kwargs)
 
-    def delete_branch_protection_rule(
-        self,
-        repository,
-        inline_object1,
-        **kwargs
-    ):
-        """delete_branch_protection_rule  # noqa: E501
-
-        This method makes a synchronous HTTP request by default. To make an
-        asynchronous HTTP request, please pass async_req=True
-
-        >>> thread = api.delete_branch_protection_rule(repository, inline_object1, async_req=True)
-        >>> result = thread.get()
-
-        Args:
-            repository (str):
-            inline_object1 (InlineObject1):
-
-        Keyword Args:
-            _return_http_data_only (bool): response data without head status
-                code and headers. Default is True.
-            _preload_content (bool): if False, the urllib3.HTTPResponse object
-                will be returned without reading/decoding response data.
-                Default is True.
-            _request_timeout (int/float/tuple): timeout setting for this request. If
-                one number provided, it will be total request timeout. It can also
-                be a pair (tuple) of (connection, read) timeouts.
-                Default is None.
-            _check_input_type (bool): specifies if type checking
-                should be done one the data sent to the server.
-                Default is True.
-            _check_return_type (bool): specifies if type checking
-                should be done one the data received from the server.
-                Default is True.
-            _host_index (int/None): specifies the index of the server
-                that we want to use.
-                Default is read from the configuration.
-            async_req (bool): execute request asynchronously
-
-        Returns:
-            None
-                If the method is called asynchronously, returns the request
-                thread.
-        """
-        kwargs['async_req'] = kwargs.get(
-            'async_req', False
-        )
-        kwargs['_return_http_data_only'] = kwargs.get(
-            '_return_http_data_only', True
-        )
-        kwargs['_preload_content'] = kwargs.get(
-            '_preload_content', True
-        )
-        kwargs['_request_timeout'] = kwargs.get(
-            '_request_timeout', None
-        )
-        kwargs['_check_input_type'] = kwargs.get(
-            '_check_input_type', True
-        )
-        kwargs['_check_return_type'] = kwargs.get(
-            '_check_return_type', True
-        )
-        kwargs['_host_index'] = kwargs.get('_host_index')
-        kwargs['repository'] = \
-            repository
-        kwargs['inline_object1'] = \
-            inline_object1
-        return self.delete_branch_protection_rule_endpoint.call_with_http_info(**kwargs)
-
     def delete_repository(
         self,
         repository,
@@ -785,71 +466,6 @@ class RepositoriesApi(object):
         kwargs['repository'] = \
             repository
         return self.delete_repository_endpoint.call_with_http_info(**kwargs)
-
-    def get_branch_protection_rules(
-        self,
-        repository,
-        **kwargs
-    ):
-        """get branch protection rules  # noqa: E501
-
-        This method makes a synchronous HTTP request by default. To make an
-        asynchronous HTTP request, please pass async_req=True
-
-        >>> thread = api.get_branch_protection_rules(repository, async_req=True)
-        >>> result = thread.get()
-
-        Args:
-            repository (str):
-
-        Keyword Args:
-            _return_http_data_only (bool): response data without head status
-                code and headers. Default is True.
-            _preload_content (bool): if False, the urllib3.HTTPResponse object
-                will be returned without reading/decoding response data.
-                Default is True.
-            _request_timeout (int/float/tuple): timeout setting for this request. If
-                one number provided, it will be total request timeout. It can also
-                be a pair (tuple) of (connection, read) timeouts.
-                Default is None.
-            _check_input_type (bool): specifies if type checking
-                should be done one the data sent to the server.
-                Default is True.
-            _check_return_type (bool): specifies if type checking
-                should be done one the data received from the server.
-                Default is True.
-            _host_index (int/None): specifies the index of the server
-                that we want to use.
-                Default is read from the configuration.
-            async_req (bool): execute request asynchronously
-
-        Returns:
-            [BranchProtectionRule]
-                If the method is called asynchronously, returns the request
-                thread.
-        """
-        kwargs['async_req'] = kwargs.get(
-            'async_req', False
-        )
-        kwargs['_return_http_data_only'] = kwargs.get(
-            '_return_http_data_only', True
-        )
-        kwargs['_preload_content'] = kwargs.get(
-            '_preload_content', True
-        )
-        kwargs['_request_timeout'] = kwargs.get(
-            '_request_timeout', None
-        )
-        kwargs['_check_input_type'] = kwargs.get(
-            '_check_input_type', True
-        )
-        kwargs['_check_return_type'] = kwargs.get(
-            '_check_return_type', True
-        )
-        kwargs['_host_index'] = kwargs.get('_host_index')
-        kwargs['repository'] = \
-            repository
-        return self.get_branch_protection_rules_endpoint.call_with_http_info(**kwargs)
 
     def get_repository(
         self,
