+++ clients/python/test/test_config_api.py
@@ -24,12 +24,6 @@ class TestConfigApi(unittest.TestCase):
     def tearDown(self):
         pass
 
-    def test_get_garbage_collection_config(self):
-        """Test case for get_garbage_collection_config
-
-        """
-        pass
-
     def test_get_lake_fs_version(self):
         """Test case for get_lake_fs_version
